Rearrange network path structure (#710)
* Move internal files in order to keep the path structure used by the other apps
- src/ -> src/app

* Fix run file

* update tests import references

* Update conftest import reference
