+++ apps/network/src/app/routes/general.py
