+++ apps/network/tests/rest_api/test_choose_model_hosts.py
@@ -1,7 +1,7 @@
 from unittest.mock import patch
 
 
-@patch("src.routes.network.network_manager")
+@patch("src.app.routes.network.network_manager")
 def test_choose_encrypted_model_host_value_error(mock_network_manager, client):
     """assert that the endpoint returns a 400 for value error."""
     mock_network_manager.connected_nodes.return_value = {}
@@ -12,7 +12,7 @@ def test_choose_encrypted_model_host_value_error(mock_network_manager, client):
     assert len(result.get_json()) == 0
 
 
-@patch("src.routes.network.network_manager")
+@patch("src.app.routes.network.network_manager")
 def test_choose_encrypted_model_host_internal_server_error(
     mock_network_manager, client
 ):
