+++ apps/network/tests/rest_api/test_delete_nodes.py
@@ -15,7 +15,7 @@ def test_delete_grid_node_bad_request_key_error(client):
     assert result.get_json().get("message") == ("Invalid JSON format.")
 
 
-@patch("src.routes.network.network_manager")
+@patch("src.app.routes.network.network_manager")
 def test_delete_grid_node_not_registered(mock_network_manager, client):
     """assert that the endpoint returns a 400 for a node not yet registered."""
     mock_network_manager.delete_node.return_value = False
@@ -30,7 +30,7 @@ def test_delete_grid_node_not_registered(mock_network_manager, client):
     )
 
 
-@patch("src.routes.network.network_manager")
+@patch("src.app.routes.network.network_manager")
 def test_delete_grid_node_internal_server_error(mock_network_manager, client):
     """assert that the endpoint returns a 500 for an unknown error
     condition."""
