+++ apps/network/tests/rest_api/test_search_encrypted_model.py
@@ -1,7 +1,7 @@
 from unittest.mock import patch
 
 
-@patch("src.routes.network.network_manager")
+@patch("src.app.routes.network.network_manager")
 def test_search_encrypted_model_bad_request_value_error(mock_network_manager, client):
     """assert that the endpoint returns a 400 for malformed JSON."""
     mock_network_manager.connected_nodes.side_effect = ValueError()
@@ -11,7 +11,7 @@ def test_search_encrypted_model_bad_request_value_error(mock_network_manager, cl
     assert result.get_json().get("message") == ("Invalid JSON format.")
 
 
-@patch("src.routes.network.network_manager")
+@patch("src.app.routes.network.network_manager")
 def test_search_encrypted_model_bad_request_key_error(mock_network_manager, client):
     """assert that the endpoint returns a 400 for missing keys in sent JSON."""
     mock_network_manager.connected_nodes.side_effect = KeyError()
@@ -21,7 +21,7 @@ def test_search_encrypted_model_bad_request_key_error(mock_network_manager, clie
     assert result.get_json().get("message") == ("Invalid JSON format.")
 
 
-@patch("src.routes.network.network_manager")
+@patch("src.app.routes.network.network_manager")
 def test_search_encrypted_model_internal_server_error(mock_network_manager, client):
     """assert that the endpoint returns a 500 for an unknown error
     condition."""
